feat: add supabase seed script
