+++ package.json
@@ -12,8 +12,9 @@
     "db:migrate": "prisma migrate dev",
     "db:migrate:deploy": "prisma migrate deploy",
     "db:reset": "prisma migrate reset",
-    "db:seed": "tsx prisma/seed/index.ts",
-    "db:setup": "prisma db push && prisma generate && tsx prisma/seed/index.ts",
+    "db:seed": "node prisma/seed/supabase-seed.ts",
+    "db:setup": "prisma db push && prisma generate && node prisma/seed/supabase-seed.ts",
+    "supabase:seed": "node prisma/seed/supabase-seed.ts",
     "db:studio": "prisma studio",
     "deploy:vercel": "vercel --prod",
     "deploy:build": "npm run build",
@@ -29,6 +30,7 @@
     "@mdxeditor/editor": "^3.39.1",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^6.11.1",
+    "@supabase/supabase-js": "^2.0.0",
     "@radix-ui/react-accordion": "^1.2.11",
     "@radix-ui/react-alert-dialog": "^1.1.14",
     "@radix-ui/react-aspect-ratio": "^1.1.7",
